backend/requirements.txt
aiofiles==23.2.1
annotated-types==0.7.0
anyio==3.7.1
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
decorator==5.2.1
distro==1.9.0
fastapi==0.104.1
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.25.2
idna==3.11
ImageIO==2.37.3
imageio-ffmpeg==0.6.0
jiter==0.13.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==1.0.0
langchain-core==1.0.0
langchain-openai==1.0.0
langgraph==1.0.10
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.11
langsmith==0.7.18
moviepy==2.2.1
numpy==2.4.3
openai==2.28.0
opencv-python==4.13.0.92
orjson==3.11.7
ormsgpack==1.12.2
packaging==25.0
pillow==10.4.0
proglog==0.1.12
pydantic==2.12.5
pydantic_core==2.41.5
pydantic-settings==2.6.1
SQLAlchemy[asyncio]==2.0.36
asyncpg==0.30.0
bcrypt==4.2.0
python-jose[cryptography]==3.3.0
pytest==8.3.4
pytest-asyncio==0.24.0
pydub==0.25.1
python-dotenv==1.0.0
python-multipart==0.0.22
PyYAML==6.0.3
regex==2026.2.28
requests==2.32.5
requests-toolbelt==1.0.0
sniffio==1.3.1
sse-starlette==1.8.2
starlette==0.27.0
tenacity==9.1.4
tiktoken==0.12.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.24.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
xxhash==3.6.0
zstandard==0.25.0
